+++ app/views/comfy/admin/cms/files/index.html.haml
@@ -5,6 +5,7 @@
   session_id = request.session_options[:id]
   session_id_value = session_id.respond_to?(:public_id) ? session_id.public_id : session_id.to_s
   cms_uploader_data = {
+    controller: 'file-upload',
     cms_uploader: {
       url: comfy_admin_cms_site_files_path(@site, source: :plupload, categories: params[:categories]),
       token: { name: request_forgery_protection_token, value: form_authenticity_token },
